Added a target to build a standalone static library for MultiMarkdown.
I am working on an app that is going to be using MultiMarkdown, and I
need lower-level access to it than the CLI can give. The Library target
can be linked in by another app and all the parsing functions can be
used directly. In particular, this allows retrieving the entire
traversable parse tree instead of converted content.